move print-level-order to k-tree
