--- challenges/print-level-order/README.md
@@ -1,12 +1,0 @@
-# print_level_order
-CF 401 Data Structures Whiteboard Challenge
-
-## Challenge
-Write a function to generate a level-order printout of a k-ary tree. That is, given a tree, display the node values, with new lines between each level of the tree.
-
-Write a function based on the specifications above, which takes in a k-ary tree. Without utilizing any of the built-in methods available to your language, return a string that contains a listing of all values in the tree, with new lines in-between each level of the tree.
-
-Write at least three test assertions for each method that you define.Ensure your tests are passing before you submit your solution.
-
-## Solution
-![whiteboard](../../assets/print-level-order.jpg)
