+++ data_structures/k_tree/README.md
@@ -15,3 +15,13 @@
         - This class should be aware of a breadth-first traversal method
         - This class should have the ability to insert a new node into the tree at a given parent node
 - Ensure that your class and any subsequent methods are properly tested, and that your test coverage is above 80%.
+
+
+### Whiteboard Challenge - Print Level Order
+Write a function to generate a level-order printout of a k-ary tree. That is, given a tree, display the node values, with new lines between each level of the tree.
+
+Write a function based on the specifications above, which takes in a k-ary tree. Without utilizing any of the built-in methods available to your language, return a string that contains a listing of all values in the tree, with new lines in-between each level of the tree.
+
+Write at least three test assertions for each method that you define.Ensure your tests are passing before you submit your solution.
+
+![whiteboard](../../assets/print-level-order.jpg)
